pyproject.toml
[project]
name = "ai-kb"
version = "0.5.0"
description = "AI 知识库 — 自动化技术情报收集与分析系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "langgraph>=0.2.60",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "hooks"]
addopts = "-q --tb=short"
markers = [
    "non_llm: deterministic tests with mocked or avoided LLM calls",
    "llm_e2e: real-provider end-to-end workflow verification",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
